+++ mvmcts/statistics/thres_greedy_statistic.h
@@ -69,7 +69,7 @@ class ThresGreedyStatistic : public UctStatistic<ThresGreedyStatistic> {
           } else if (b.second.action_count_ == 0) {
             return false;
           } else {
-            return (ThresholdComparator<Eigen::VectorXf>(thr))(
+            return (ThresholdComparator<Eigen::VectorXd>(thr))(
                 a.second.action_value_, b.second.action_value_);
           }
         });
